backend/requirements.txt
django
djangorestframework
pillow
dj_database_url
psycopg2-binary
requests
markdown
